- Fixed a few compiler warning (size_t/int) - VStudio solution file is now compiling - Defined variables for the Windows redistributables (makes the script version independ from VStudio)

+++ base/MainWindow.cpp
@@ -62,7 +62,7 @@ void MainWindow::Init()
 	MainWindowBase::Init();
 
 	// init tournament classes (if there are none present)
-	for (int i(0); i < m_pCategoryManager->CategoryCount(); ++i)
+	for (int i(0); i < (int)m_pCategoryManager->CategoryCount(); ++i)
 	{
 		FightCategory t("");
 		m_pCategoryManager->GetCategory(i, t);
@@ -86,7 +86,7 @@ void MainWindow::on_actionManageClasses_triggered()
 
 		m_pUi->comboBox_weight_class->clear();
 
-		for (int i(0); i < m_pCategoryManager->CategoryCount(); ++i)
+		for (int i(0); i < (int)m_pCategoryManager->CategoryCount(); ++i)
 		{
 			FightCategory t("");
 			m_pCategoryManager->GetCategory(i, t);
